requirements.txt
dspy-ai>=3.1.3
google-genai>=1.64.0
python-dotenv>=1.2.1
streamlit==1.42.0
pandas==2.2.0